pyproject.toml
[build-system]
requires = ["setuptools>=77.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nameko-keycloak"
description = "Helpers to integrate Single Sign-On in nameko-based applications using Keycloak."
version = "2.1.0"
requires-python = ">=3.9"
authors = [{name="Emplocity"}]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Operating System :: Unix",
    "Operating System :: POSIX",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Utilities",
]
readme = "README.rst"
license = " Apache-2.0"
license-files = ["LICENSE"]
urls = {Homepage = "https://github.com/emplocity/nameko-keycloak"}
dependencies = [
    "nameko>=2",
    "jwcrypto>=1.5",
    "python-keycloak>=4.2",
    "werkzeug>=1.0",
]

[project.optional-dependencies]
test = ["pytest>=8"]

[tool.setuptools]
include-package-data = true
zip-safe = false

[tool.mypy]
python_version = "3.9"
mypy_path = ["src/", "tests/"]
namespace_packages = true
no_implicit_optional = true
no_implicit_reexport = true
strict_equality = true
warn_redundant_casts = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "eventlet.*",
    "jwcrypto.*",
    "keycloak.*",
    "nameko.*",
    "pytest.*",
    "werkzeug.*",
]
ignore_missing_imports = true

[tool.tox]
legacy_tox_ini = """
[tox]
isolated_build = True
envlist =
    check, py{39,310,311,312,313}-nameko{2,3}

[testenv]
deps =
    nameko2: nameko>=2.14,<3
    nameko3: nameko>=3.0.0rc11,<4
    -rtest_requirements.txt
commands =
    {posargs:pytest --cov={envsitepackagesdir}/nameko_keycloak -vv}

[testenv:check]
deps =
    -rtest_requirements.txt
    build
    check-manifest
    docutils
    pygments
    readme-renderer
    twine
skip_install = true
commands =
    python -m build
    twine check dist/*
    check-manifest {toxinidir}
    flake8 src tests
    isort --verbose --check-only --diff src tests
    mypy src

[gh-actions]
python =
    3.9: py39-nameko2, py39-nameko3
    3.10: py310-nameko2, py310-nameko3
    3.11: py311-nameko2, py311-nameko3
    3.12: py312-nameko2, py312-nameko3
    3.13: py313-nameko2, py313-nameko3, check

"""
